refactor(server): extract recipe + run-audit + templates routes
Slice 6 of the server.ts split — final big slice. Lifts 16 routes
covering the recipe authoring loop (CRUD + lint + run), the run-audit
log (/runs, /runs/:seq, replay, plan), the public template registry
(/templates, with its 5-min cache), recipe installation
(/recipes/install), and /activation-metrics out of server.ts and into
src/recipeRoutes.ts.

DI shape: RecipeRouteDeps interface with 12 nullable callbacks the
bridge wires onto Server post-construction (runRecipeFn, recipesFn,
runReplayFn, etc.). Server.ts dispatches via:

    if (
      tryHandleRecipeRoute(req, res, parsedUrl, {
        recipesFn: this.recipesFn,
        // ... 11 more
      })
    ) return;

Cache lifetime: the /templates 5-min GitHub registry cache used to live
as private _templatesCache/_templatesCacheTs fields on Server. Lifetime
is process-wide either way (Server is a singleton in practice), so
hoisting to module-level state in recipeRoutes.ts is equivalent and
avoids threading a mutable holder through deps.

server.ts: 2107 → 1551 lines (-26% this slice). Cumulative across all
six slice PRs (#97, #98, #99, #100, #101, this): server.ts is now
**-2028 lines (-57%)** from where it started this session (3579 →
1551). 18 connectors, 7 OAuth callbacks, 2 inbox routes, 6 OAuth 2.0
endpoints, 2 MCP-card routes, and 16 recipe/run routes are all now in
focused modules.

Bonus cleanup along the way:
- Removed now-unused private fields _templatesCache, _templatesCacheTs
- Removed unused imports: os, path, computeActivationSummary,
  loadActivationMetrics

Verified:
- Full bridge test suite: 323 files / 4500 tests green (no regressions)
- Includes the 8 server-recipes-content.test.ts tests + the run/replay/
  plan integration tests in webhookRecipes.test.ts and friends — all
  pass against the extracted module
- Zero LSP diagnostics on both files (errors + warnings)
- Biome clean
