+++ src/server.ts
@@ -1,12 +1,7 @@
 import { EventEmitter } from "node:events";
 import http from "node:http";
-import os from "node:os";
-import path from "node:path";
 import { WebSocket, WebSocketServer as WsServer } from "ws";
-import {
-  computeSummary as computeActivationSummary,
-  loadMetrics as loadActivationMetrics,
-} from "./activationMetrics.js";
+import type { ActivityListener } from "./activityTypes.js";
 import { handleApprovalsStream, routeApprovalRequest } from "./approvalHttp.js";
 import { getApprovalQueue } from "./approvalQueue.js";
 import { saveBridgeConfigDriver } from "./config.js";
@@ -27,14 +22,13 @@ import {
   defaultConfigPath as patchworkConfigPath,
   saveConfig as savePatchworkConfig,
 } from "./patchworkConfig.js";
+import { tryHandleRecipeRoute } from "./recipeRoutes.js";
+import type { RecipeDraft } from "./recipesHttp.js";
 import { PACKAGE_VERSION } from "./version.js";
 
 const LOOPBACK_HOSTS = new Set(["localhost", "127.0.0.1", "[::1]"]);
 const SESSION_ID_RE =
   /^[0-9a-f]{8}-[0-9a-f]{4}-[0-9a-f]{4}-[0-9a-f]{4}-[0-9a-f]{12}$/i;
-
-import type { ActivityListener } from "./activityTypes.js";
-import type { RecipeDraft } from "./recipesHttp.js";
 
 interface AliveWebSocket extends WebSocket {
   isAlive: boolean;
@@ -264,8 +258,6 @@ export class Server extends EventEmitter<ServerEvents> {
     | null = null;
   /** Patchwork: set by bridge to list active agent sessions for the dashboard. */
   public sessionsFn: (() => SessionSummary[]) | null = null;
-  private _templatesCache: unknown = null;
-  private _templatesCacheTs = 0;
   /** Patchwork: set by bridge to answer GET /sessions/:id with per-session event stream + approvals. */
   public sessionDetailFn:
     | ((id: string) => {
@@ -847,574 +839,23 @@ export class Server extends EventEmitter<ServerEvents> {
         return;
       }
 
-      const recipeNameRunMatch =
-        req.method === "POST"
-          ? /^\/recipes\/([^/]+)\/run$/.exec(parsedUrl.pathname)
-          : null;
-      if (recipeNameRunMatch) {
-        const nameFromPath = decodeURIComponent(recipeNameRunMatch[1] ?? "");
-        const chunks: Buffer[] = [];
-        req.on("data", (c: Buffer) => chunks.push(c));
-        req.on("end", () => {
-          void (async () => {
-            try {
-              const body = Buffer.concat(chunks).toString("utf-8");
-              const parsed = body
-                ? (JSON.parse(body) as {
-                    vars?: Record<string, string>;
-                    inputs?: Record<string, string>;
-                  })
-                : {};
-              const varsRaw = parsed.vars ?? parsed.inputs;
-              const vars =
-                varsRaw &&
-                typeof varsRaw === "object" &&
-                !Array.isArray(varsRaw)
-                  ? (varsRaw as Record<string, string>)
-                  : undefined;
-              if (!this.runRecipeFn) {
-                res.writeHead(503, { "Content-Type": "application/json" });
-                res.end(
-                  JSON.stringify({
-                    ok: false,
-                    error:
-                      "Recipe execution unavailable — requires --claude-driver subprocess",
-                  }),
-                );
-                return;
-              }
-              const result = await this.runRecipeFn(nameFromPath, vars);
-              res.writeHead(result.ok ? 200 : 400, {
-                "Content-Type": "application/json",
-              });
-              res.end(JSON.stringify(result));
-            } catch {
-              res.writeHead(400, { "Content-Type": "application/json" });
-              res.end(
-                JSON.stringify({ ok: false, error: "Invalid JSON body" }),
-              );
-            }
-          })();
-        });
-        return;
-      }
-      if (parsedUrl.pathname === "/recipes/run" && req.method === "POST") {
-        const chunks: Buffer[] = [];
-        req.on("data", (c: Buffer) => chunks.push(c));
-        req.on("end", () => {
-          void (async () => {
-            try {
-              const body = Buffer.concat(chunks).toString("utf-8");
-              const parsed = JSON.parse(body || "{}") as {
-                name?: string;
-                vars?: Record<string, string>;
-              };
-              const name = parsed.name;
-              const vars =
-                parsed.vars &&
-                typeof parsed.vars === "object" &&
-                !Array.isArray(parsed.vars)
-                  ? (parsed.vars as Record<string, string>)
-                  : undefined;
-              if (typeof name !== "string" || !name) {
-                res.writeHead(400, { "Content-Type": "application/json" });
-                res.end(JSON.stringify({ ok: false, error: "name required" }));
-                return;
-              }
-              if (!this.runRecipeFn) {
-                res.writeHead(503, { "Content-Type": "application/json" });
-                res.end(
-                  JSON.stringify({
-                    ok: false,
-                    error:
-                      "Recipe execution unavailable — requires --claude-driver subprocess",
-                  }),
-                );
-                return;
-              }
-              const result = await this.runRecipeFn(name, vars);
-              res.writeHead(result.ok ? 200 : 400, {
-                "Content-Type": "application/json",
-              });
-              res.end(JSON.stringify(result));
-            } catch {
-              res.writeHead(400, { "Content-Type": "application/json" });
-              res.end(
-                JSON.stringify({ ok: false, error: "Invalid JSON body" }),
-              );
-            }
-          })();
-        });
-        return;
-      }
+      // ── Recipe / runs / templates routes (extracted to src/recipeRoutes.ts) ─
       if (
-        parsedUrl.pathname === "/activation-metrics" &&
-        req.method === "GET"
+        tryHandleRecipeRoute(req, res, parsedUrl, {
+          recipesFn: this.recipesFn,
+          loadRecipeContentFn: this.loadRecipeContentFn,
+          saveRecipeContentFn: this.saveRecipeContentFn,
+          deleteRecipeContentFn: this.deleteRecipeContentFn,
+          lintRecipeContentFn: this.lintRecipeContentFn,
+          saveRecipeFn: this.saveRecipeFn,
+          setRecipeEnabledFn: this.setRecipeEnabledFn,
+          runsFn: this.runsFn,
+          runDetailFn: this.runDetailFn,
+          runPlanFn: this.runPlanFn,
+          runReplayFn: this.runReplayFn,
+          runRecipeFn: this.runRecipeFn,
+        })
       ) {
-        try {
-          const metrics = loadActivationMetrics();
-          const summary = computeActivationSummary(metrics);
-          res.writeHead(200, { "Content-Type": "application/json" });
-          res.end(JSON.stringify({ metrics, summary }));
-        } catch (err) {
-          res.writeHead(500, { "Content-Type": "application/json" });
-          res.end(
-            JSON.stringify({
-              error: err instanceof Error ? err.message : String(err),
-            }),
-          );
-        }
-        return;
-      }
-      if (parsedUrl.pathname === "/runs" && req.method === "GET") {
-        try {
-          const sp = parsedUrl.searchParams;
-          const limitRaw = sp.get("limit");
-          const afterRaw = sp.get("after");
-          const trigger = sp.get("trigger");
-          const status = sp.get("status");
-          const recipe = sp.get("recipe");
-          const limit = limitRaw ? Number.parseInt(limitRaw, 10) : Number.NaN;
-          const after = afterRaw ? Number.parseInt(afterRaw, 10) : Number.NaN;
-          const runs =
-            this.runsFn?.({
-              ...(Number.isFinite(limit) && { limit }),
-              ...(trigger && { trigger }),
-              ...(status && { status }),
-              ...(recipe && { recipe }),
-              ...(Number.isFinite(after) && { after }),
-            }) ?? [];
-          res.writeHead(200, { "Content-Type": "application/json" });
-          res.end(JSON.stringify({ runs }));
-        } catch (err) {
-          res.writeHead(500, { "Content-Type": "application/json" });
-          res.end(
-            JSON.stringify({
-              error: err instanceof Error ? err.message : String(err),
-            }),
-          );
-        }
-        return;
-      }
-      // GET /runs/:seq — single run detail (includes stepResults if present)
-      const runDetailMatch =
-        req.method === "GET"
-          ? /^\/runs\/(\d+)$/.exec(parsedUrl.pathname)
-          : null;
-      if (runDetailMatch?.[1]) {
-        const seq = Number.parseInt(runDetailMatch[1], 10);
-        try {
-          const run = this.runDetailFn?.(seq) ?? null;
-          if (!run) {
-            res.writeHead(404, { "Content-Type": "application/json" });
-            res.end(JSON.stringify({ error: "not_found" }));
-          } else {
-            res.writeHead(200, { "Content-Type": "application/json" });
-            res.end(JSON.stringify({ run }));
-          }
-        } catch (err) {
-          res.writeHead(500, { "Content-Type": "application/json" });
-          res.end(
-            JSON.stringify({
-              error: err instanceof Error ? err.message : String(err),
-            }),
-          );
-        }
-        return;
-      }
-      // POST /runs/:seq/replay — VD-4 mocked replay. Re-runs the recipe
-      // with all tool/agent execution intercepted to return captured
-      // outputs from the original run. No external IO, no side effects.
-      // Real-mode replay is not exposed here yet — must ship separately
-      // with confirmation UX + kill-switch interaction.
-      const runReplayMatch =
-        req.method === "POST"
-          ? /^\/runs\/(\d+)\/replay$/.exec(parsedUrl.pathname)
-          : null;
-      if (runReplayMatch?.[1]) {
-        const seq = Number.parseInt(runReplayMatch[1], 10);
-        try {
-          if (!this.runReplayFn) {
-            res.writeHead(503, { "Content-Type": "application/json" });
-            res.end(JSON.stringify({ error: "replay_unavailable" }));
-            return;
-          }
-          const result = await this.runReplayFn(seq);
-          if (result.error === "run_not_found") {
-            res.writeHead(404, { "Content-Type": "application/json" });
-          } else if (!result.ok) {
-            res.writeHead(500, { "Content-Type": "application/json" });
-          } else {
-            res.writeHead(200, { "Content-Type": "application/json" });
-          }
-          res.end(JSON.stringify(result));
-        } catch (err) {
-          const msg = err instanceof Error ? err.message : String(err);
-          res.writeHead(500, { "Content-Type": "application/json" });
-          res.end(JSON.stringify({ error: msg }));
-        }
-        return;
-      }
-      // GET /runs/:seq/plan — dry-run plan for the recipe that produced this run
-      const runPlanMatch =
-        req.method === "GET"
-          ? /^\/runs\/(\d+)\/plan$/.exec(parsedUrl.pathname)
-          : null;
-      if (runPlanMatch?.[1]) {
-        const seq = Number.parseInt(runPlanMatch[1], 10);
-        try {
-          const run = this.runDetailFn?.(seq) ?? null;
-          if (!run) {
-            res.writeHead(404, { "Content-Type": "application/json" });
-            res.end(JSON.stringify({ error: "run_not_found" }));
-            return;
-          }
-          if (!this.runPlanFn) {
-            res.writeHead(503, { "Content-Type": "application/json" });
-            res.end(JSON.stringify({ error: "plan_unavailable" }));
-            return;
-          }
-          // triggerSource appends ":agent" suffix — strip before file lookup
-          const recipeName = (run.recipeName as string).replace(/:agent$/, "");
-          const plan = await this.runPlanFn(recipeName);
-          res.writeHead(200, { "Content-Type": "application/json" });
-          res.end(JSON.stringify({ plan }));
-        } catch (err) {
-          const msg = err instanceof Error ? err.message : String(err);
-          const status =
-            msg.includes("not found") || msg.includes("ENOENT") ? 404 : 500;
-          res.writeHead(status, { "Content-Type": "application/json" });
-          res.end(JSON.stringify({ error: msg }));
-        }
-        return;
-      }
-      if (req.url === "/recipes" && req.method === "POST") {
-        const chunks: Buffer[] = [];
-        req.on("data", (c: Buffer) => chunks.push(c));
-        req.on("end", () => {
-          try {
-            const body = Buffer.concat(chunks).toString("utf-8");
-            const draft = JSON.parse(body || "{}") as RecipeDraft;
-            if (typeof draft.name !== "string" || !draft.name) {
-              res.writeHead(400, { "Content-Type": "application/json" });
-              res.end(JSON.stringify({ ok: false, error: "name required" }));
-              return;
-            }
-            if (!this.saveRecipeFn) {
-              res.writeHead(503, { "Content-Type": "application/json" });
-              res.end(
-                JSON.stringify({
-                  ok: false,
-                  error: "Recipe saving unavailable",
-                }),
-              );
-              return;
-            }
-            const result = this.saveRecipeFn(draft);
-            res.writeHead(result.ok ? 201 : 400, {
-              "Content-Type": "application/json",
-            });
-            res.end(JSON.stringify(result));
-          } catch {
-            res.writeHead(400, { "Content-Type": "application/json" });
-            res.end(JSON.stringify({ ok: false, error: "Invalid JSON body" }));
-          }
-        });
-        return;
-      }
-      const recipePatchMatch = /^\/recipes\/([^/]+)$/.exec(parsedUrl.pathname);
-      if (recipePatchMatch && req.method === "PATCH") {
-        const name = decodeURIComponent(recipePatchMatch[1] ?? "");
-        const chunks: Buffer[] = [];
-        req.on("data", (c: Buffer) => chunks.push(c));
-        req.on("end", () => {
-          try {
-            const body = JSON.parse(
-              Buffer.concat(chunks).toString("utf-8"),
-            ) as { enabled?: boolean };
-            if (typeof body.enabled !== "boolean") {
-              res.writeHead(400, { "Content-Type": "application/json" });
-              res.end(
-                JSON.stringify({
-                  ok: false,
-                  error: "enabled (boolean) required",
-                }),
-              );
-              return;
-            }
-            if (!this.setRecipeEnabledFn) {
-              res.writeHead(503, { "Content-Type": "application/json" });
-              res.end(JSON.stringify({ ok: false, error: "Not available" }));
-              return;
-            }
-            const result = this.setRecipeEnabledFn(name, body.enabled);
-            res.writeHead(result.ok ? 200 : 400, {
-              "Content-Type": "application/json",
-            });
-            res.end(JSON.stringify(result));
-          } catch {
-            res.writeHead(400, { "Content-Type": "application/json" });
-            res.end(JSON.stringify({ ok: false, error: "Invalid JSON" }));
-          }
-        });
-        return;
-      }
-      if (parsedUrl.pathname === "/recipes/lint" && req.method === "POST") {
-        const chunks: Buffer[] = [];
-        req.on("data", (c: Buffer) => chunks.push(c));
-        req.on("end", () => {
-          try {
-            const body = JSON.parse(
-              Buffer.concat(chunks).toString("utf-8"),
-            ) as {
-              content?: string;
-            };
-            if (typeof body?.content !== "string") {
-              res.writeHead(400, { "Content-Type": "application/json" });
-              res.end(
-                JSON.stringify({
-                  ok: false,
-                  error: "content (string) required",
-                }),
-              );
-              return;
-            }
-            if (!this.lintRecipeContentFn) {
-              res.writeHead(503, { "Content-Type": "application/json" });
-              res.end(
-                JSON.stringify({
-                  ok: false,
-                  error: "Recipe lint unavailable",
-                }),
-              );
-              return;
-            }
-            const result = this.lintRecipeContentFn(body.content);
-            res.writeHead(200, { "Content-Type": "application/json" });
-            res.end(JSON.stringify(result));
-          } catch {
-            res.writeHead(400, { "Content-Type": "application/json" });
-            res.end(JSON.stringify({ ok: false, error: "Invalid JSON body" }));
-          }
-        });
-        return;
-      }
-      const recipeContentMatch = /^\/recipes\/([^/]+)$/.exec(
-        parsedUrl.pathname,
-      );
-      if (recipeContentMatch && req.method === "GET") {
-        const name = decodeURIComponent(recipeContentMatch[1] ?? "");
-        if (!this.loadRecipeContentFn) {
-          res.writeHead(503, { "Content-Type": "application/json" });
-          res.end(
-            JSON.stringify({ ok: false, error: "Recipe content unavailable" }),
-          );
-          return;
-        }
-        const result = this.loadRecipeContentFn(name);
-        if (!result) {
-          res.writeHead(404, { "Content-Type": "application/json" });
-          res.end(JSON.stringify({ ok: false, error: "Recipe not found" }));
-          return;
-        }
-        res.writeHead(200, { "Content-Type": "application/json" });
-        res.end(JSON.stringify(result));
-        return;
-      }
-      if (recipeContentMatch && req.method === "PUT") {
-        const name = decodeURIComponent(recipeContentMatch[1] ?? "");
-        const chunks: Buffer[] = [];
-        req.on("data", (c: Buffer) => chunks.push(c));
-        req.on("end", () => {
-          try {
-            const body = JSON.parse(
-              Buffer.concat(chunks).toString("utf-8"),
-            ) as { content?: string };
-            if (typeof body.content !== "string") {
-              res.writeHead(400, { "Content-Type": "application/json" });
-              res.end(
-                JSON.stringify({
-                  ok: false,
-                  error: "content (string) required",
-                }),
-              );
-              return;
-            }
-            if (!this.saveRecipeContentFn) {
-              res.writeHead(503, { "Content-Type": "application/json" });
-              res.end(
-                JSON.stringify({
-                  ok: false,
-                  error: "Recipe content saving unavailable",
-                }),
-              );
-              return;
-            }
-            const result = this.saveRecipeContentFn(name, body.content);
-            res.writeHead(result.ok ? 200 : 400, {
-              "Content-Type": "application/json",
-            });
-            res.end(JSON.stringify(result));
-          } catch {
-            res.writeHead(400, { "Content-Type": "application/json" });
-            res.end(JSON.stringify({ ok: false, error: "Invalid JSON body" }));
-          }
-        });
-        return;
-      }
-      if (recipeContentMatch && req.method === "DELETE") {
-        const name = decodeURIComponent(recipeContentMatch[1] ?? "");
-        if (!this.deleteRecipeContentFn) {
-          res.writeHead(503, { "Content-Type": "application/json" });
-          res.end(
-            JSON.stringify({
-              ok: false,
-              error: "Recipe deletion unavailable",
-            }),
-          );
-          return;
-        }
-        const result = this.deleteRecipeContentFn(name);
-        const status = result.ok
-          ? 200
-          : result.error === "Recipe not found"
-            ? 404
-            : 400;
-        res.writeHead(status, { "Content-Type": "application/json" });
-        res.end(JSON.stringify(result));
-        return;
-      }
-      if (req.url === "/recipes" && req.method === "GET") {
-        try {
-          const data = this.recipesFn?.() ?? { recipesDir: null, recipes: [] };
-          res.writeHead(200, { "Content-Type": "application/json" });
-          res.end(JSON.stringify(data));
-        } catch (err) {
-          res.writeHead(500, { "Content-Type": "application/json" });
-          res.end(
-            JSON.stringify({
-              error: err instanceof Error ? err.message : String(err),
-            }),
-          );
-        }
-        return;
-      }
-      if (parsedUrl.pathname === "/templates" && req.method === "GET") {
-        try {
-          const now = Date.now();
-          if (
-            !this._templatesCache ||
-            now - this._templatesCacheTs > 5 * 60 * 1000
-          ) {
-            const ghRes = await fetch(
-              "https://raw.githubusercontent.com/patchworkos/recipes/main/index.json",
-            );
-            if (!ghRes.ok) {
-              throw new Error(`GitHub returned ${ghRes.status}`);
-            }
-            this._templatesCache = (await ghRes.json()) as unknown;
-            this._templatesCacheTs = now;
-          }
-          res.writeHead(200, { "Content-Type": "application/json" });
-          res.end(JSON.stringify(this._templatesCache));
-        } catch (err) {
-          res.writeHead(502, { "Content-Type": "application/json" });
-          res.end(
-            JSON.stringify({
-              ok: false,
-              error: err instanceof Error ? err.message : String(err),
-            }),
-          );
-        }
-        return;
-      }
-      if (parsedUrl.pathname === "/recipes/install" && req.method === "POST") {
-        let body = "";
-        req.on("data", (chunk: Buffer) => {
-          body += chunk.toString();
-        });
-        req.on("end", async () => {
-          try {
-            const { source } = JSON.parse(body) as { source?: string };
-            if (!source) {
-              res.writeHead(400, { "Content-Type": "application/json" });
-              res.end(
-                JSON.stringify({ ok: false, error: "Missing source field" }),
-              );
-              return;
-            }
-            const githubPrefix = "github:patchworkos/recipes/recipes/";
-            let fetchUrl: string;
-            let recipeName: string;
-            if (source.startsWith(githubPrefix)) {
-              recipeName = source.slice(githubPrefix.length);
-              fetchUrl = `https://raw.githubusercontent.com/patchworkos/recipes/main/recipes/${recipeName}/${recipeName}.yaml`;
-            } else if (source.startsWith("https://")) {
-              fetchUrl = source;
-              const urlParts = fetchUrl.split("/");
-              recipeName = (urlParts[urlParts.length - 1] ?? "recipe").replace(
-                /\.ya?ml$/i,
-                "",
-              );
-            } else {
-              res.writeHead(400, { "Content-Type": "application/json" });
-              res.end(
-                JSON.stringify({
-                  ok: false,
-                  error: "Unsupported source format",
-                }),
-              );
-              return;
-            }
-            const yamlRes = await fetch(fetchUrl);
-            if (!yamlRes.ok) {
-              throw new Error(
-                `Fetch failed: ${yamlRes.status} ${yamlRes.statusText}`,
-              );
-            }
-            const yamlText = await yamlRes.text();
-            const tmpFile = path.join(
-              os.tmpdir(),
-              `patchwork-install-${Date.now()}-${recipeName}.yaml`,
-            );
-            const { writeFileSync, mkdirSync, unlinkSync } = await import(
-              "node:fs"
-            );
-            writeFileSync(tmpFile, yamlText, "utf-8");
-            let result: { action: "created" | "replaced"; name: string };
-            try {
-              const recipesDir = path.join(
-                os.homedir(),
-                ".patchwork",
-                "recipes",
-              );
-              mkdirSync(recipesDir, { recursive: true });
-              const { installRecipeFromFile } = await import(
-                "./recipes/installer.js"
-              );
-              const installResult = installRecipeFromFile(tmpFile, {
-                recipesDir,
-              });
-              result = { action: installResult.action, name: recipeName };
-            } finally {
-              try {
-                unlinkSync(tmpFile);
-              } catch {
-                // best-effort cleanup
-              }
-            }
-            res.writeHead(200, { "Content-Type": "application/json" });
-            res.end(JSON.stringify({ ok: true, ...result }));
-          } catch (err) {
-            res.writeHead(500, { "Content-Type": "application/json" });
-            res.end(
-              JSON.stringify({
-                ok: false,
-                error: err instanceof Error ? err.message : String(err),
-              }),
-            );
-          }
-        });
         return;
       }
       const sessionDetailMatch = /^\/sessions\/([A-Za-z0-9-]+)$/.exec(
